shell() expects node to be specified with argument :n, not :d
puppet_module_install helper fails when attempting to copy a puppet
module to a non-default node (in multi-node test environments).

The helper creates a directory for the module via shell("mkdir...")
before using scp to copy module files. Currently it passes the node
using wrong parameter (:d instead of :n), so this command is always
executed on the default node and the following scp() fails since the
directory does not exist.